utm_partner_binding/pyproject.toml
[project]
name = "utm-partner-binding"
version = "0.1.0"
description = "Сервис автоматической привязки партнеров к сделкам и лидам по UTM меткам"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fast-bitrix24>=1.8.11",
    "loguru==0.7.3",
    "python-dotenv==1.2.1",
    "fastapi",
    "uvicorn",
]

